Remove: Delete EMD processor script and related API functions
- Removed the `emd_processor.py` script, which provided functionality for processing EMD data and generating Excel files.
- Deleted associated API functions for processing EMD data, spelling numbers, and validating EMD data, streamlining the codebase.
